refactor: 컴포넌트 내의 함수 및 상수 분리
