+++ lib/date.ts
@@ -1,0 +1,10 @@
+/**
+ * 날짜 문자열을 "N days ago" 형식으로 변환합니다.
+ */
+export function timeAgo(dateStr: string): string {
+  const diff = Date.now() - new Date(dateStr).getTime();
+  const days = Math.floor(diff / 86400000);
+  if (days === 0) return "today";
+  if (days === 1) return "1 day ago";
+  return `${days} days ago`;
+}
